fix headers rule switch instanceof

+++ src/main/java/net/pms/image/ExifInfo.java
@@ -1,3 +1,20 @@
+/*
+ * This file is part of Universal Media Server, based on PS3 Media Server.
+ *
+ * This program is a free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; version 2
+ * of the License only.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
 package net.pms.image;
 
 import java.awt.color.ColorSpace;
@@ -261,7 +278,7 @@ public abstract class ExifInfo extends ImageInfo {
 				) {
 					Integer i = ((ExifDirectoryBase) directory).getInteger(ExifDirectoryBase.TAG_COLOR_SPACE);
 					if (i != null) {
-						((ExifParseInfo) parsedInfo).exifColorSpace = ExifColorSpace.typeOf(i.intValue());
+						((ExifParseInfo) parsedInfo).exifColorSpace = ExifColorSpace.typeOf(i);
 					}
 				}
 			} else if (directory instanceof ExifThumbnailDirectory exifThumbnailDirectory &&
@@ -295,7 +312,7 @@ public abstract class ExifInfo extends ImageInfo {
 	 *         if unknown.
 	 */
 	public int getExifVersion() {
-		return exifVersion != null ? exifVersion.intValue() : UNKNOWN;
+		return exifVersion != null ? exifVersion : UNKNOWN;
 	}
 
 	/**
